refactor(linter): Move gitlab-ci linters to new file
